implement TimerIntentService to work around possible broadcast receiver timeouts
