+++ AndroidManifest.xml
@@ -275,6 +275,8 @@
             <action android:name="android.service.chooser.ChooserTargetService" />
         </intent-filter>
     </service>
+    <service android:name=".connect.TimerIntentService"
+            android:permission="android.permission.BIND_JOB_SERVICE" />
 
     <receiver android:name=".connect.TimerReceiver" android:enabled="true"/>
